refactor(sct_config): extract types, helpers and defaults into modules
Splits the module-level parts of the 5.4K-line config out of config.py:

- types.py     custom Annotated aliases, their BeforeValidator converters,
               IgnoredType/InputType and the SctField descriptor
- helpers.py   appendable-option merging, env sub-key parsing, region
               counting and the docker-image defaults cache
- defaults.py  backend lists and the eight per-backend requirement tables
- __init__.py  the public re-exports

Three things are not pure moves:

- is_config_option_appendable() read SCTConfiguration.model_fields, which
  would make helpers import config. It and merge_dicts_append_strings()
  now take the model class as a parameter instead.
- The instance catalog directory was derived from __file__, which resolves
  to sdcm/data/instance_catalog once the file sits a directory deeper. The
  resulting FileNotFoundError is swallowed with a warning, so constraint-
  based sizing would have silently stopped resolving on every run. It now
  uses sct_abs_path(), with a regression test that fails on the old form.
- Both loggers keep the name "sdcm.sct_config" rather than __name__, so log
  lines and caplog targets do not move.

The class-level tables stay class fields assigned from defaults.py
constants: Pydantic deep-copies mutable defaults per instance, and
_check_backend_defaults mutates them.

__init__.py deliberately re-exports only what the repo imports. Names that
config.py merely imports for its own use are left out so that a stale
mock.patch("sdcm.sct_config.<name>") raises AttributeError instead of
silently patching an attribute no call site reads.

Field count and order, the generated docs and the full config dump for a
docker test case are all unchanged.

Part of SCT-525.
